+++ mooc-core/src/main/java/cn/edu/lingnan/mooc/core/repository/CourseRepository.java
@@ -14,7 +14,7 @@ import java.util.List;
  * @see JpaRepository 支持常用增删查改
  * @see JpaSpecificationExecutor 支持多条件分页
  */
-public interface CourseRepository extends JpaRepository<Course, Integer>,JpaSpecificationExecutor<Course> {
+public interface CourseRepository extends JpaRepository<Course, Long>,JpaSpecificationExecutor<Course> {
 
     /**
      * 根据标签id查询课程
@@ -24,6 +24,6 @@ public interface CourseRepository extends JpaRepository<Course, Integer>,JpaSpec
      */
     @Query(value="select c.* from course c,course_tag_rel r where c.id = r.course_id and c.status=1 and r.tag_id in (?1) group by c.id"
             ,countQuery = "select count(c.id) from course c,course_tag_rel r where c.id = r.course_id and c.status=1 and r.tag_id in (?1) group by c.id",nativeQuery=true)
-    Page<Course> findCourseByTagList(List<Integer> tagIdList, Pageable pageable);
+    Page<Course> findCourseByTagList(List<Long> tagIdList, Pageable pageable);
 
 }
